# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "잰잰 API 문서",
    "description": "미션, 리뷰 및 가게 관련 API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Missions",
      "description": "미션 관련 API"
    },
    {
      "name": "Reviews",
      "description": "리뷰 관련 API"
    },
    {
      "name": "Stores",
      "description": "가게 관련 API"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/missions/start": {
      "post": {
        "tags": ["Missions"],
        "description": "특정 가게에서 미션을 시작합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "missionId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/missions/ongoing/{storeId}": {
      "get": {
        "tags": ["Stores"],
        "description": "특정 가게의 진행 중인 미션 목록을 조회합니다.",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/missions/complete/{storeId}/{missionId}": {
      "patch": {
        "tags": ["Missions"],
        "description": "특정 가게에서 지정된 미션을 완료 상태로 변경합니다.",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "missionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/reviews/add": {
      "post": {
        "tags": ["Reviews"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "storeId": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/stores/{storeId}/missions": {
      "get": {
        "tags": ["Stores"],
        "description": "특정 가게에 대한 미션 목록을 조회합니다.",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}